Hashtag following bug
Fixed bug where when user was logged out and toggled follow on a
hashtag, and the proceeded to sign in on an account already following
that hashtag, there would be an error.

Also, fixes bug where if user would toggle a hashtag and then go to the
same hashtag, but in a different case (e.g. testHAsh and TesTHASh),
there would throw an error.