Comment’s swipe to edit/delete feature is now determined by permissions, rather than client-side logic.
The result is a sort of all or nothing situation, where either all
comments in the view will be editable/deletable or none of them will.

If we want some particular comments to editable/deletable (like one’s
posted by the current user), we would have to either (a) inject some
client-side code, or (b) change the way permissions are being handled,
so that they are being sent on a per comment basis.